ざっくりとクラス分割
